backend/requirements.txt
alembic==1.17.2
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
APScheduler==3.11.2
bcrypt==4.0.1
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
cssselect==1.3.0
cssutils==2.11.1
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
emails==0.6
fastapi==0.128.0
greenlet==3.3.0
h11==0.16.0
httptools==0.7.1
idna==3.11
iniconfig==2.3.0
Jinja2==3.1.6
lxml==6.0.2
Mako==1.3.10
MarkupSafe==3.0.3
more-itertools==10.8.0
packaging==25.0
passlib==1.7.4
pillow==12.1.0
pluggy==1.6.0
premailer==3.10.0
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
Pygments==2.19.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.21
PyYAML==6.0.3
reportlab==4.4.7
requests==2.32.5
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sse-starlette==3.1.2
starlette==0.50.0
tenacity==9.1.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uvicorn==0.40.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
slowapi==0.1.9
httpx==0.28.1
openpyxl
